Create component - CategoryButton
