+++ src/pages/index.tsx
@@ -14,9 +14,7 @@ import { FlexContainerBetween, FlexContainerAlignCenter } from 'src/styles/FlexC
 import { TABLET_MIN_WIDTH } from 'src/models/constants'
 import { sleep } from 'src/utils/commons'
 import useGoToPage from 'src/hooks/useGoToPage'
-import { Theme, makeStyles, createStyles } from '@material-ui/core/styles'
-import ButtonBase from '@material-ui/core/ButtonBase'
-import Typography from '@material-ui/core/Typography'
+import CategoryButton from 'src/components/CategoryButton'
 
 const PADDING_TOP = '3rem'
 
@@ -48,94 +46,6 @@ const GridContainerUl = styled.ul<{ onlyImage: boolean }>`
 const ClickableDiv = styled.div`
   cursor: pointer;
 `
-const images = [
-  {
-    url: 'https://cdn.pixabay.com/photo/2017/05/04/21/23/cupcakes-2285209_1280.jpg',
-    title: '카테고리별 보기',
-    width: '50%',
-  },
-  {
-    url:
-      'https://search.pstatic.net/common/?src=http%3A%2F%2Fimage.nmv.naver.net%2Fblog_2020_08_25_2801%2Ff114664d-e68b-11ea-a019-246e963a49a8_01.jpg&type=sc960_832',
-    title: '테마별 보기',
-    width: '50%',
-  },
-]
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      display: 'flex',
-      flexWrap: 'wrap',
-      minWidth: 300,
-      width: '100%',
-    },
-    image: {
-      position: 'relative',
-      height: 200,
-      [theme.breakpoints.down('xs')]: {
-        width: '100% !important', // Overrides inline-style
-        height: 100,
-      },
-      '&:hover, &$focusVisible': {
-        zIndex: 1,
-        '& $imageBackdrop': {
-          opacity: 0.15,
-        },
-        '& $imageMarked': {
-          opacity: 0,
-        },
-        '& $imageTitle': {
-          border: '4px solid currentColor',
-        },
-      },
-    },
-    focusVisible: {},
-    imageButton: {
-      position: 'absolute',
-      left: 0,
-      right: 0,
-      top: 0,
-      bottom: 0,
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      color: theme.palette.common.white,
-    },
-    imageSrc: {
-      position: 'absolute',
-      left: 0,
-      right: 0,
-      top: 0,
-      bottom: 0,
-      backgroundSize: 'cover',
-      backgroundPosition: 'center 40%',
-    },
-    imageBackdrop: {
-      position: 'absolute',
-      left: 0,
-      right: 0,
-      top: 0,
-      bottom: 0,
-      backgroundColor: theme.palette.common.black,
-      opacity: 0.4,
-      transition: theme.transitions.create('opacity'),
-    },
-    imageTitle: {
-      position: 'relative',
-      padding: `${theme.spacing(2)}px ${theme.spacing(4)}px ${theme.spacing(1) + 6}px`,
-    },
-    imageMarked: {
-      height: 3,
-      width: 18,
-      backgroundColor: theme.palette.common.white,
-      position: 'absolute',
-      bottom: -2,
-      left: 'calc(50% - 9px)',
-      transition: theme.transitions.create('opacity'),
-    },
-  })
-)
 
 function HomePage() {
   const [isLoadingMenus, setIsLoadingMenus] = useState(false)
@@ -143,8 +53,6 @@ function HomePage() {
   const [onlyImage, toggleOnlyImage] = useBoolean(false)
 
   const goToSearchPage = useGoToPage('/search')
-
-  const classes = useStyles()
 
   async function fetchMoreMenus() {
     setIsLoadingMenus(true)
@@ -184,39 +92,8 @@ function HomePage() {
             layout="fill"
           />
         </ImageRatioWrapper>
-        <div className={classes.root}>
-          {images.map((image) => (
-            <ButtonBase
-              focusRipple
-              key={image.title}
-              className={classes.image}
-              focusVisibleClassName={classes.focusVisible}
-              style={{
-                width: image.width,
-              }}
-            >
-              <span
-                className={classes.imageSrc}
-                style={{
-                  backgroundImage: `url(${image.url})`,
-                }}
-              />
-              <span className={classes.imageBackdrop} />
-              <span className={classes.imageButton}>
-                <Typography
-                  component="span"
-                  variant="subtitle1"
-                  color="inherit"
-                  className={classes.imageTitle}
-                >
-                  {image.title}
-                  <span className={classes.imageMarked} />
-                </Typography>
-              </span>
-            </ButtonBase>
-          ))}
-        </div>
 
+        <CategoryButton></CategoryButton>
         <div>정렬 기준</div>
         <button onClick={toggleOnlyImage}>사진만 보기</button>
